% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{BackwardProbs}
\alias{BackwardProbs}
\title{Calculate Backward Probabilities from a Backward Table Object}
\usage{
BackwardProbs(bck, log = TRUE)
}
\arguments{
\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{log}{logical; if TRUE (default), the backward probabilities p are returned as log(p).}
}
\value{
matrix of backward probabilities
}
\description{
Provides an easy function for calculating the log backward probabilities from a backward table object bck.
}
\details{
The (i,j)-th element of of the returned matrix is the probability that j copies i at locus fwd$l and the haplotypes
observed from locus fwd$l+1 up through locus L.
}
\examples{
bck <- MakeBackwardTable()
Backward(bck, 100, Pi, mu, rho)
BackwardProbs(bck)
}
\seealso{
\code{\link{MakeBackwardTable}} to generate backward table;
  \code{\link{Backward}} to propogate a backward table to a new locus.
}
