% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheAllHaplotypes.R
\name{CacheAllHaplotypesH5}
\alias{CacheAllHaplotypesH5}
\title{Create memory cache of haplotypes}
\usage{
CacheAllHaplotypesH5(hdf5.file, transpose = FALSE)
}
\arguments{
\item{hdf5.file}{the name of the file which the haplotypes are to be read from.}

\item{transpose}{a logical value indicating whether to switch the
interpretation of the slowest changing dimension (hence switching the
number of haplotypes and the length)}
}
\value{
Nothing is returned by the function.  However, a status message is
  output indicating how the dimensions of the HDF5 \code{haps} object have
  been interpreted.  If this message shows the dimensions incorrectly ordered,
  call the function again with argument \code{transpose = TRUE}.
}
\description{
Load haplotypes from hard drive to memory.
}
\details{
To achieve higher performance, kalis internally represents haplotypes
in an efficient raw binary format in memory.  This function will load haplotypes
from an HDF5 file and convert this into kalis' internal format ready for use
by the other functions in this package.

The package expects a 2-dimensional object named \code{haps} at the root
level of the HDF5 file.  Haplotypes should be stored in the slowest changing
dimension as defined in the HDF5 specification (note that different languages
treat this as rows or columns).  If the haplotypes are stored in the other
dimension then simply set the argument \code{transpose = TRUE}.
If the user is unsure of the convention of
the language they used to create the HDF5 file, then the simplest approach is
to just load the data specifying only the HDF5 file name and then confirm
that number of haplotypes and their length have not been exchanged.
}
\examples{
# Examples
\dontrun{
# Load haplotypes to cache from an HDF5 file on disk
CacheAllHaplotypesH5("myhaps.h5")

# If the diagnostic message printed during the above indicates the numbers
# of haplotypes and their length are the wrong way around, reload with
# argument to transpose
CacheAllHaplotypesH5("myhaps.h5", transpose = TRUE)

# When correct orientation is known, can avoid diagnostic messages for running
# in script files
suppressMessages(CacheAllHaplotypesH5("myhaps.h5"))
}

}
\seealso{
\code{\link{ClearHaplotypeCache}} to remove the haplotypes from
  the cache and free the memory.
}
