% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheAllSequences.R
\name{CacheAllSequencesH5}
\alias{CacheAllSequencesH5}
\title{Create memory cache of haplotype sequences}
\usage{
CacheAllSequencesH5(hdf5.file, transpose = FALSE)
}
\arguments{
\item{from_recipient}{first recipient haplotype if creating a partial forward
table.  By default includes from the first recipient haplotype.}

\item{to_recipient}{last recipient haplotype if creating a partial forward
table.  By default includes to the last recipient haplotype.}
}
\value{
Nothing is returned by the function.
}
\description{
Load haplotype sequences from hard drive to memory.
}
\details{
To achieve higher performance, kalis internally represents haplotype sequences
in an efficient raw binary format in memory.  This function will load sequences
from the hard drive and convert into this format ready for use by the other
functions in this package.

The package assumes haplotypes are stored in the slowest changing dimension
as per the HDF5 specification (note that different languages treat this as
rows or columns).  If the user is unsure of the convention of the language
they used to create the HDF5 file, then the simplest approach is probably to
test loading the data and confirm that number of haplotypes and sequence
length have not been exchanged.
}
\examples{
# Examples
\dontrun{
# Create a forward table for the hidden Markov model incorporating all
# recipient and donor haplotypes
fwd <- MakeForwardTable()

# Create a forward table for the hidden Markov model incorporating only
# recipient haplotypes 100 to 200 (inclusive) and all donor haplotypes.
fwd <- MakeForwardTable(100, 200)

# This table is uninitialised, but ready to pass to the Forward function
# which will trigger initialisation and propagation from the first locus.
# For example, initialise and propagate forward to locus 10:
Forward(fwd, 10, morgan.dist, Ne, gamma, mu, nthreads = 8)
}

}
\seealso{
\code{\link{Forward}} to propagate the newly created table forward
  through the genome.
}
