% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheSummary.R
\name{CacheSummary}
\alias{CacheSummary}
\alias{N}
\alias{L}
\title{Retrieve information about the haplotype cache}
\usage{
CacheSummary()

N()

L()
}
\value{
\code{CacheSummary()} prints information about the current state of the kalis
cache.

\code{N()} returns the number of haplotypes currently in the kalis cache, or
\code{NULL} if the cache is empty.

\code{L()} returns the number of variants currently in the kalis cache, or
\code{NULL} if the cache is empty.
}
\description{
Retrieve information about the haplotype cache
}
