% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iterator.R
\name{CalcCheckpointTables}
\alias{CalcCheckpointTables}
\title{Calculate Checkpoint Tables}
\usage{
CalcCheckpointTables(propagation.cost, max.num.checkpoints, use.R = FALSE)
}
\arguments{
\item{propagation.cost}{a non-negative vector such that \code{propagation.cost[i]} gives the relative amount of time or cost required to propagate \code{i} steps}

\item{max.num.checkpoints}{the maximum number of checkpoints that should be considered when building the checkpoint table.}

\item{use.R}{a logical, when TRUE use base R rather than C implementation of table building. Defaults to FALSE.}
}
\value{
a list containing:
\describe{
\item{\code{cost}}{the matrix \eqn{F} in Christ et al. (2024)}
\item{\code{index}}{the matrix \eqn{H} in Christ et al. (2024)}
}
}
\description{
Calculate look up tables for solving optimal checkpointing problems with dynamic programming.
}
\references{
Christ, R., Wang, X., Aslett, L.J.M., Steinsaltz, D. and Hall, I. (2024) "Clade Distillation for Genome-wide Association Studies." bioRxiv 2024.09.30.615852. Available at: \doi{10.1101/2024.09.30.615852}.
}
