% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CladeMat.R
\name{CladeMat}
\alias{CladeMat}
\title{Fast clade matrix construction}
\usage{
CladeMat(
  fwd,
  bck,
  M,
  unit.dist,
  thresh = 0.2,
  max1var = FALSE,
  nthreads = min(parallel::detectCores(logical = FALSE), fwd$to_recipient -
    fwd$from_recipient + 1)
)
}
\arguments{
\item{fwd}{a \code{kalisForwardTable} object, as returned by \code{\link[=MakeForwardTable]{MakeForwardTable()}} and propagated to a target variant by \code{\link[=Forward]{Forward()}}.
This table must be at the same variant location as argument \code{bck}.}

\item{bck}{a \code{kalisBackwardTable} object, as returned by \code{\link[=MakeBackwardTable]{MakeBackwardTable()}} and propagated to a target variant by \code{\link[=Backward]{Backward()}}.
This table must be at the same variant location as argument \code{fwd}.}

\item{M}{a matrix with half the number of rows and columns as the corresponding forward/backward tables.
This matrix is overwritten in place with the clade matrix result for performance reasons.}

\item{unit.dist}{the change in distance that is expected to correspond to a single mutation (typically \eqn{-\log(\mu)}) for the LS model)}

\item{thresh}{a regularization parameter: \if{html}{\out{<!-- TODO: can't parse this sentence -->}} differences distances must exceed this threshold (in \code{unit.dist} units) in order to used in forming the local relatedness matrix.
Defaults to \code{0.2}.}

\item{max1var}{a logical regularization parameter.
When \code{TRUE}, differences in distances exceeding 1 \code{unit.dist} are set to 1 (so that any edge in the latent ancestral tree with multiple mutations on them are treated as if only one mutation was on it).}

\item{nthreads}{the number of CPU cores to use.
By default uses the \code{parallel} package to detect the number of physical cores.}
}
\value{
A list, the first element contains a list of tied nearest neighbours (one for each haplotype).
Other elements of the returned list are for internal use by \code{\link[=PruneCladeMat]{PruneCladeMat()}} to allow for efficient removal of singletons and sprigs.
}
\description{
Constructs a clade matrix using forward and backward tables.
The clade matrix captures genetic relatedness information in the distances from the Li & Stephens model that are not captured in the called clades.
}
\details{
\code{CladeMat()} uses the forward and backward tables to construct the corresponding clade matrix which can then be tested, for example using a standard quadratic form score statistic.
}
\examples{
# TODO


}
\references{
Christ, R., Wang, X., Aslett, L.J.M., Steinsaltz, D. and Hall, I. (2024) "Clade Distillation for Genome-wide Association Studies." bioRxiv 2024.09.30.615852. Available at: \doi{10.1101/2024.09.30.615852}.
}
