% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clades.R
\name{CladeMat_old}
\alias{CladeMat_old}
\title{CladeMat OLD}
\usage{
CladeMat_old(
  x,
  ploidy = 2L,
  sprigs.to.prune = NULL,
  assemble = TRUE,
  use.forking = FALSE,
  forking.chunk.size = 100L,
  mc.preschedule = FALSE,
  nthreads = 1L
)
}
\arguments{
\item{x}{a \code{kalisClades} object returned by \code{\link[=Clades]{Clades()}}}

\item{ploidy}{an integer, the ploidy of the organism}

\item{sprigs.to.prune}{a \code{kalisSprigs} object returned by \code{\link[=Sprigs]{Sprigs()}} encoding sprigs that should be excluded from the matrix returned}

\item{assemble}{a logical, if \code{FALSE} return the clade matrix as a list of columns rather than as a symmetrized matrix}

\item{use.forking}{a logical, should forked processes be used?}

\item{nthreads}{the number of CPU cores to use. Currently, no parallelism is used.}
}
\value{
a matrix representation of the probabilistic clades provided
}
\description{
Utility for contructing a probabilistic clade matrix
}
