% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clades.R
\name{Clades}
\alias{Clades}
\title{Probabilistic Clades}
\usage{
Clades(
  fwd,
  bck,
  pars,
  beta.theta.opts = NULL,
  safety.checks = FALSE,
  neighbors = FALSE,
  use.forking = FALSE,
  forking.chunk.size = 100L,
  mc.preschedule = FALSE,
  nthreads = 1L
)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{pars}{a \code{kalisParameters} object, as returned by
\code{\link{Parameters}}.}

\item{beta.theta.opts}{a list; see Details for \code{\link{DistMat}}.}

\item{safety.checks}{a logical, should safety checks be applied to the distances?  See \code{\link{DistMat}}.}

\item{neighbors}{a logical, should nearest neighbors be pre-calculated?  See \code{\link{Neighbors}}.}

\item{use.forking}{a logical, should forked processes be used?}

\item{nthreads}{the number of CPU cores to use. Currently, no parallelism is used.}
}
\value{
a \code{kalisClades} object encoding probabilistic clade calls
}
\description{
Utility for calling probabilistic clades at, in between, or excluding variants.
}
