% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheHaplotypes.R
\name{ClearHaplotypeCache}
\alias{ClearHaplotypeCache}
\title{Remove all cached haplotypes and free memory}
\usage{
ClearHaplotypeCache()
}
\value{
Nothing is returned.
}
\description{
Remove all the haplotypes that were cached by a previous caching call and free up the memory that was allocated for future use.
}
\details{
To achieve higher performance, kalis internally represents haplotypes in an efficient raw binary format in memory which cannot be directly viewed or manipulated in R, though you can extract a view from it using \code{\link[=QueryCache]{QueryCache()}}.
In particular, this cache sits outside R's memory management and will never be garbage collected (unless R is quit or the package is unloaded).
Therefore, this function is provided to enable freeing the memory used by this cache.
}
\examples{
# First fill the cache with the toy data included in the package
data("SmallHaps")
CacheHaplotypes(SmallHaps)

# Verify it is there
CacheSummary()

# Now clear
ClearHaplotypeCache()

# Verify it is gone
CacheSummary()

}
\references{
Aslett, L.J.M. and Christ, R.R. (2024) "kalis: a modern implementation of the Li & Stephens model for local ancestry inference in R", \emph{BMC Bioinformatics}, \strong{25}(1). Available at: \doi{10.1186/s12859-024-05688-8}.
}
\seealso{
\code{\link[=CacheHaplotypes]{CacheHaplotypes()}} to create a haplotype cache;
\code{\link[=QueryCache]{QueryCache()}} to view the cache contents in an R matrix.
}
