% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheAllSequences.R
\name{ClearSequenceCache}
\alias{ClearSequenceCache}
\title{Remove all cached sequences and free memory}
\usage{
ClearSequenceCache()
}
\value{
Nothing is returned.
}
\description{
Remove all the sequences which were cached using a previous caching call and
free the memory that was allocated up for future use.
}
\details{
To achieve higher performance, kalis internally represents haplotype sequences
in an efficient raw binary format in memory which cannot be directly viewed
or manipulated in R.  In particular, this cache sits outside R's memory
management and will never be garbage collected (unless R is quit or the
package is unloaded).  Therefore, this function is provided to enable freeing
the memory used by this cache.
}
\examples{
# Examples
\dontrun{
}

}
\seealso{
\code{\link{Forward}} to propagate the newly created table forward
  through the genome.
}
