% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dedip.R
\name{DedipPar}
\alias{DedipPar}
\title{Dedip in Parallel}
\usage{
DedipPar(
  M,
  x = rep(1/nrow(M[[1]]), nrow(M[[1]])),
  fwd,
  bck,
  pars,
  fwd_rho = NULL,
  bck_rho = NULL,
  from_recipient = 1,
  nthreads = 1
)
}
\description{
Dedip in Parallel
}
