% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{ForwardProbs}
\alias{ForwardProbs}
\title{Calculate Forward Probabilities from a Forward Table Object}
\usage{
ForwardProbs(fwd, log = TRUE)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{log}{logical; if TRUE (default), the forward probabilities p are returned as log(p).}
}
\value{
matrix of forward probabilities
}
\description{
Provides an easy function for calculating the log forward probabilities from a forward table object fwd.
}
\details{
The (i,j)-th element of of the returned matrix is the probability that j copies i at locus fwd$l and the haplotypes
observed from locus 1 up through locus fwd$l.
}
\examples{
fwd <- MakeForwardTable()
Forward(fwd, 100, Pi, mu, rho)
ForwardProbs(fwd)
}
\seealso{
\code{\link{MakeForwardTable}} to generate forward table;
  \code{\link{Forward}} to propogate a forward table to a new locus.
}
