% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableCache.R
\name{ForwardUsingTableCache}
\alias{ForwardUsingTableCache}
\title{Title}
\usage{
ForwardUsingTableCache(fwd, pars, cache, t = fwd$l + 1, nthreads = 1)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{cache}{a cache of forward tables as generated by
\code{\link{CreateForwardTableCache}}}

\item{t}{a locus position to move the forward table to, using the cache for
performance.  Must be greater than or equal to locus position of table
provided in \code{fwd}.}

\item{nthreads}{the number of CPU cores on which to run}

\item{Pi}{a matrix of background copying probabilities.  Can also provide a
scalar value for uniform background copying probability.}

\item{mu}{a vector of ...}

\item{rho}{a vector of ...}
}
\value{
There is nothing returned.  For performance reasons, the forward
  table which was passed in is updated in-place.
}
\description{
Short description
}
\details{
Detailed description
}
\examples{
\dontrun{
ForwardUsingTableCache(...)
}

}
\seealso{
\code{\link{MakeForwardTable}} to make a forward table;

  \code{\link{CreateForwardTableCache}} to generate a cache;

  \code{\link{Forward}} for forward function without using a cache.
}
