% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clades.R
\name{Neighbors}
\alias{Neighbors}
\title{Neighbors}
\usage{
Neighbors(x, use.forking = FALSE, nthreads = 1L)
}
\arguments{
\item{x}{a \code{kalisClades} object returned by \code{\link[=Clades]{Clades()}}}

\item{use.forking}{a logical, should forked processes be used?}

\item{nthreads}{the number of CPU cores to use. Currently, no parallelism is used.}
}
\value{
a \code{kalisNeighbors} encoding the nearest neighbors for each recipient haplotype
}
\description{
Utility for calling tied nearest neighbors for each recipient haplotype
}
