% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{Parameters}
\alias{Parameters}
\title{Setup kalis HMM parameters}
\usage{
Parameters(rho = rep(0, get("L", envir = pkgVars) - 1), mu = 1e-08,
  Pi = NULL)
}
\arguments{
\item{rho}{recombination probability vector (must be L-1 long), see \code{CalcRho}}

\item{mu}{a scalar (for uniform) or vector (for varying) mutation costs.}

\item{Pi}{leaving the default of uniform copying probabilities is recommended for
computational efficiency.  If desired, a full matrix of background copying
probabilities can be provided, such that the (i,j)-th element is the background
probability that j copies i.  Hence, (a) the diagonal must be zero; and (b)
the columns of Pi must sum to 1.}
}
\value{
A new parameters environment.
}
\description{
Sets up the genetics parameters to be used for a problem.
}
\details{
Detailed description
}
\examples{
\dontrun{
pars <- Parameters(...)
}

}
\seealso{
\code{\link{MakeForwardTable}}, \code{\link{MakeForwardTable}} which
construct table objects which internally reference a parameters environment.
}
