% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{PostProbs}
\alias{PostProbs}
\title{Calculate Posterior Marginal Probabilities from a Forward Table Object and a Backward Table Object}
\usage{
PostProbs(fwd, bck, log = FALSE)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{log}{logical; if FALSE, the posterior marginal probabilities p are returned as log(p).}
}
\value{
matrix of posterior marginal probabilities
}
\description{
Provides an easy function for calculating the posterior marginal probabilities.
}
\details{
The forward and backward tables must be at the same locus in order for them to be combined to yield the posterior marginal probabilities at locus l.
The (j,i)-th element of of the returned matrix is the probability that j is copied by i at the current locus of the two tables, l, given the haplotypes observed (from locus 1 to L).
Note: each column represents an independent HMM.

By convention, every diagonal element is zero.
}
\examples{
\dontrun{
PostProbs(fwd,bck)
}

}
\seealso{
\code{\link{DistMat}} to generate calculate \eqn{d_{ji}} distances directly
\code{\link{MakeForwardTable}} to generate Forward table;
\code{\link{Forward}} to propogate a Backward table to a new locus;
\code{\link{MakeBackwardTable}} to generate a Backward table;
\code{\link{Backward}} to propogate a Backward table to a new locus.
}
