% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheAllHaplotypes.R
\name{QueryCache}
\alias{QueryCache}
\title{Retrieve haplotypes from memory cache}
\usage{
QueryCache(ids = NA, start = 1, length = NA)
}
\arguments{
\item{ids}{which haplotypes to retrieve from the cache}

\item{start}{the first locus position to retrieve for the specified haplotypes.
Defaults to the beginning of the haplotype.}

\item{length}{the number of consecutive loci to return from the start position.
Defaults to length of entire haplotype from \code{start} argument to end.}
}
\value{
A matrix of 0/1 integers with \code{length} rows and
\code{length(ids)} columns, such that haplotypes appear in columns.
}
\description{
Retrieve haplotypes from the memory cache, converting the raw binary into
a simple integer 0/1 vector for use in R.
}
\details{
To achieve higher performance, kalis internally represents haplotypes
in an efficient raw binary format in memory which cannot be directly viewed
or manipulated in R.  This function enables users to copy whole or partial
haplotypes out of this low-level format and into a standard R
vector of 0's and 1's.
}
\examples{
# Examples
\dontrun{
QueryCache(...)
}

}
\seealso{
\code{\link{CacheHaplotypes}} to fill the memory cache with
haplotypes.
}
