% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker.R
\name{ResetTable}
\alias{ResetTable}
\title{Reset Forward/Backward tables}
\usage{
ResetTable(tbl)
}
\arguments{
\item{tbl}{a \code{kalisForwardTable} or \code{kalisBackwardTable} object
which is to be reset.}
}
\description{
Resets the provided forward/backward table so it can propagate from the start/end of the Li and Stephens hidden Markov model.
}
\details{
It will usually be faster to reset a forward/backward table rather than remove and make a new one.  This function marks a table as reset so that it will be propagated as if freshly allocated.
}
\examples{
# Examples
\dontrun{
# Create a forward table for the hidden Markov model incorporating all
# recipient and donor haplotypes
fwd <- MakeForwardTable()

# Propagate forward to variant 10:
Forward(fwd, pars, 10, nthreads = 8)

# This does **NOT** work as intended, cannot reverse from variant 10:
Forward(fwd, pars, 5, nthreads = 8)

# Could reset the table though:
ResetTable(fwd)
Forward(fwd, pars, 5, nthreads = 8)
}

}
\seealso{
\code{\link{MakeForwardTable}}, \code{\link{MakeBackwardTable}} to
create fresh tables.
}
