% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clades.R
\name{Sprigs_old}
\alias{Sprigs_old}
\title{Sprigs}
\usage{
Sprigs_old(x, use.forking = FALSE, nthreads = 1L, add.self = TRUE)
}
\arguments{
\item{x}{a \code{kalisNeighbors} object returned by \code{\link[=Neighbors]{Neighbors()}}, a \code{kalisClades} object returned by \code{\link[=Clades]{Clades()}} with \code{neighbors = TRUE}, or a list}

\item{use.forking}{a logical, should forked processes be used?}

\item{nthreads}{the number of CPU cores to use. Currently, no parallelism is used.}
}
\value{
a \code{kalisSprigs} object assigning each haplotype to a sprig
}
\description{
Utility for calling sprigs from probabilistic clades
}
