% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FB.R
\name{Backward}
\alias{Backward}
\title{Propagate an HMM backward table}
\usage{
Backward(bck, t, Pi, mu, rho, nthreads = 1)
}
\arguments{
\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{t}{a locus position to move the backward table to.  Must be less than
or equal to locus position of table provided in \code{bck}.}

\item{Pi}{a matrix of background copying probabilities.  Can also provide a
scalar value for uniform background copying probability.}

\item{mu}{a vector of ...}

\item{rho}{a vector of ...}

\item{nthreads}{the number of CPU cores on which to run}
}
\value{
There is nothing returned.  For performance reasons, the backward
  table which was passed in is updated in-place.
}
\description{
Takes a backward table and propagates it in-place to an earlier locus
position.
}
\details{
Detailed description
}
\examples{
bck <- MakeBackwardTable()
Backward(bck, 100, Pi, mu, rho)
}
\seealso{
\code{\link{MakeBackwardTable}} to generate backward table;
  \code{\link{Forward}} for analagous forward induction function.
}
