% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FB.R
\name{Backward}
\alias{Backward}
\title{Propagate an HMM backward table}
\usage{
Backward(bck, pars, t = bck$l - 1, nthreads = 1)
}
\arguments{
\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{t}{a locus position to move the backward table to.  Must be less than
or equal to locus position of table provided in \code{bck}.}

\item{nthreads}{the number of CPU cores on which to run.}

\item{morgan.dist}{a vector of recombination distances between loci, in Morgans.
Note element i of this vector should be the distance between loci i and i+1
(not i and i-1), and thus length one less than the haplotype length.}

\item{Ne}{a scalar for the effective population size.  Can be autotuned, see ...}

\item{gamma}{a scalar power to which the Morgan distances are raised.  Can be
autotuned, see ...}

\item{mu}{a scalar (for uniform) or vector (for varying) mutation costs.}

\item{Pi}{leaving the default of uniform copying probabilities is recommended for
computational efficiency.  If desired, a full matrix of background copying
probabilities can be provided, such that the (i,j)-th element is the background
probability that j copies i.  Hence, (a) the diagonal must be zero; and (b)
the columns of Pi must sum to 1.}
}
\value{
There is nothing returned.  For performance reasons, the backward
  table which was passed in is updated in-place.
}
\description{
Takes a backward table and propagates it in-place to an earlier locus
position.
}
\details{
Detailed description
}
\examples{
\dontrun{
bck <- MakeBackwardTable()
Backward(bck, 100, Pi, mu, rho)
}

}
\seealso{
\code{\link{MakeBackwardTable}} to generate backward table;
  \code{\link{Forward}} for analagous forward induction function.
}
