% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{BackwardProbs}
\alias{BackwardProbs}
\title{Calculate Normalized Backward Probabilities from a Backward Table Object}
\usage{
BackwardProbs(bck, log = FALSE)
}
\arguments{
\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{log}{logical; if TRUE, the backward probabilities p are returned as log(p).}
}
\value{
matrix of backward probabilities
}
\description{
Provides an easy function for calculating the normalized backward probabilities from a backward table object bck.
}
\details{
The (j,i)-th element of of the returned matrix is the probability of observing the haplotypes from locus bck$l+1
up through locus L given that j is copied by i at bck$l divided by the sum over i of those probabilities.
Although this matrix does not have a clean probabilistic interpretation, it captures how similar haplotypes are to each
other downstream of bck$l.  Note: each column represents an independent HMM.

By convention, every diagonal element is zero.
}
\examples{
\dontrun{
BackwardProbs(bck)
}

}
\seealso{
\code{\link{MakeBackwardTable}} to generate backward table;
\code{\link{Backward}} to propogate a backward table to a new locus.
}
