% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{BackwardProbs}
\alias{BackwardProbs}
\title{Calculate Backward Probabilities from a Backward Table Object}
\usage{
BackwardProbs(bck, log = FALSE)
}
\arguments{
\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{log}{logical; if TRUE, the backward probabilities p are returned as log(p).}
}
\value{
matrix of backward probabilities
}
\description{
Provides an easy function for calculating the log backward probabilities from a backward table object bck.
}
\details{
The (i,j)-th element of of the returned matrix is the probability that j copies i at locus bck$l given the haplotypes
observed from locus bck$l+1 up through locus L.  Note: each column represents the backward probabilities from an independent HMM.

Since a haplotype is not allowed to copy from itself, every diagonal element is zero.
}
\examples{
BackwardProbs(bck)
}
\seealso{
\code{\link{MakeBackwardTable}} to generate backward table;
  \code{\link{Backward}} to propogate a backward table to a new locus.
}
