% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{CalcRho}
\alias{CalcRho}
\title{Calculate recombination probabilities}
\usage{
CalcRho(cM = 0, s = 1, gamma = 1, floor = TRUE)
}
\arguments{
\item{cM}{a vector specifying the recombination distance between variants in centimorgans.
Note element i of this vector should be the distance between variants \code{i} and \code{i+1} (not \code{i} and \code{i-1}), and thus length one less than the number of variants.
This can be easily obtained by applying \code{\link[=diff]{diff()}} to a recombination map 'CDF'.
By default, recombination probabilities are zero, meaning that all variants are perfectly linked.}

\item{s}{a scalar multiplier on the recombination map (related to effective population size).}

\item{gamma}{a scalar power to which the Morgan distances are raised.}

\item{floor}{if \code{TRUE} (default) any recombination probabilities below machine precision (\code{1e-16}) will be zeroed out.
If \code{FALSE} raw recombination probabilities will be preserved.}
}
\value{
A vector of recombination probabilities which can be used as the \code{rho} argument to the \code{\link[=Parameters]{Parameters()}} function when creating a parameter set.
}
\description{
Calculate the recombination probabilities, rho, from a given recombination map for the haplotype data currently in the kalis cache.
}
\details{
This is a utility function to calculate the recombination probabilities rho (the Li and Stephens hidden Markov model transition probabilities) from a recombination map/distances.

\strong{NOTE:} the corresponding haplotype data \emph{must} have already been inserted into the kalis cache by a call to \code{\link[=CacheHaplotypes]{CacheHaplotypes()}}, since this function performs checks to confirm the dimensionality matches.

See page 3 in Supplemental Information for the original ChromoPainter paper (Lawson et al., 2012) for motivation behind our parameterisation, which is as follows:

\deqn{\rho = 1 - \exp(-s \times cM^\gamma)}{\rho = 1 - exp(-s * cM^\gamma)}

For a complete description, see the main kalis paper, Aslett and Christ (2024).
}
\examples{
# Load the mini example data and recombination map from the package built-in #' # dataset
data("SmallHaps")
data("SmallMap")

# Load haplotypes into cache
CacheHaplotypes(SmallHaps)

# Compute the recombination probabilities, leaving s and gamma as default
rho <- CalcRho(diff(SmallMap))
rho

}
\references{
Aslett, L.J.M. and Christ, R.R. (2024) "kalis: a modern implementation of the Li & Stephens model for local ancestry inference in R", \emph{BMC Bioinformatics}, \strong{25}(1). Available at: \doi{10.1186/s12859-024-05688-8}.

Lawson, D.J., Hellenthal, G., Myers, S. and Falush, D. (2012) "Inference of population structure using dense haplotype data", \emph{PLoS genetics}, \strong{8}(1). Available at: \doi{10.1371/journal.pgen.1002453}.
}
\seealso{
\code{\link[=Parameters]{Parameters()}} to use the resulting recombination probabilities to construct a \code{kalisParameters} object.
}
