% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{CalcRho}
\alias{CalcRho}
\title{Calculate recombination parameter rho from more standard genetics parameters}
\usage{
CalcRho(morgan.dist = 0, Ne = 1, gamma = 1,
  threshold.lower = 1e-16, threshold.upper = Inf)
}
\arguments{
\item{morgan.dist}{a vector specifying the recombination distance between loci in Morgans .
Note element i of this vector should be the distance between loci i and i+1
(not i and i-1), and thus length one less than the haplotype length.  Can be easily obtained by applying \code{diff} to a recombination map "CDF".}

\item{Ne}{a scalar for the effective population size.}

\item{gamma}{a scalar power to which the Morgan distances are raised.}

\item{threshold.lower}{sets the minimal threshold value for rho.  Any values initially
calculated below this value are reset to the value specified here.}

\item{threshold.upper}{sets the maximal threshold value for rho.  Any values initially
calculated above this value are reset to the value specified here.}
}
\value{
A vector of transition probabilities
}
\description{
Sets up the Li & Stephens HMM transition probabilities rho to be used for a problem.
}
\details{
See page 3 in Supplemental Information for the original ChromoPainter paper for motivation behind our parameterization:
Lawson, Hellenthal, Myers, and Falush (2012), "Inference of population structure using dense haplotype data", PLoS Genetics, 8 (e1002453).

\link{Insert link to paper for rho maths.}
}
\examples{
\dontrun{
pars <- CalcRho(...)
}
}
\seealso{
\code{\link{Parameters}} to use the resulting transition probabilities to construct a kalisParameters object
}
