% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTraces.R
\name{CalcTraces}
\alias{CalcTraces}
\title{Fast Calculation of Matrix Trace and Hilbert Schmidt Norm}
\usage{
CalcTraces(
  M,
  tX,
  tQ,
  J,
  from_recipient = 1L,
  nthreads = min(parallel::detectCores(logical = FALSE), ncol(M))
)
}
\arguments{
\item{M}{a real symmetric R matrix}

\item{tX}{\code{t((Q \%*\% (J\%*\%Q)) - (M \%*\% Q))}}

\item{tQ}{\code{t(Q)}}

\item{J}{\code{crossprod(Q, M)}}

\item{from_recipient}{haplotype index at which to start trace calculation --- useful for distributed computation (experimental feature, more documentation to come\if{html}{\out{<!-- TODO -->}})}

\item{nthreads}{the number of CPU cores to use.
By default uses the \code{parallel} package to detect the number of physical cores.}
}
\value{
A list containing three elements:

\describe{
\item{\code{trace}}{the trace, \eqn{\mathrm{tr}(PMP)};}
\item{\code{hsnorm2}}{the \emph{squared} Hilbert Schmidt Norm of \eqn{PMP}, \eqn{\mathrm{tr}((PMP)'PMP)};}
\item{\code{diag}}{the diagonal of \eqn{PMP}.}
}
}
\description{
Provides multithreaded calculation of trace and Hilbert Schmidt Norm of a matrix \eqn{PMP} (where \eqn{P} is a projection matrix and \eqn{M} is real symmetric) without explicitly forming \eqn{PMP}.
}
\details{
\eqn{P} here is assumed to have the form \eqn{I-QQ'} for some matrix \eqn{Q} of orthogonal columns.
}
\examples{
# TODO

}
