% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTraces.R
\name{CalcTraces}
\alias{CalcTraces}
\title{Title TODO}
\usage{
CalcTraces(
  M,
  tX,
  tQ,
  J,
  from_recipient = 1L,
  nthreads = min(parallel::detectCores(logical = FALSE), ncol(M))
)
}
\arguments{
\item{M}{TODO}

\item{tX}{TODO}

\item{tQ}{TODO}

\item{J}{TODO}

\item{from_recipient}{TODO}

\item{nthreads}{the number of CPU cores to use.
By default uses the \code{parallel} package to detect the number of physical cores.}
}
\value{
TODO
}
\description{
Short Description TODO
}
\details{
Long Description TODO
}
\examples{
# TODO

}
