% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CladeMat.R
\name{CladeMat}
\alias{CladeMat}
\title{Fast Clade Matrix Construction}
\usage{
CladeMat(
  fwd,
  bck,
  M,
  unit.dist,
  thresh = 0.2,
  max1var = FALSE,
  nthreads = min(parallel::detectCores(logical = FALSE), fwd$to_recipient -
    fwd$from_recipient + 1)
)
}
\description{
Fast Clade Matrix Construction
}
