% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clades.R
\name{Clades}
\alias{Clades}
\title{Probabilistic Clades}
\usage{
Clades(
  fwd,
  bck,
  pars,
  beta.theta.opts = NULL,
  safety.checks = FALSE,
  neighbors = FALSE,
  use.forking = FALSE,
  forking.chunk.size = 100L,
  mc.preschedule = FALSE,
  nthreads = 1L
)
}
\arguments{
\item{fwd}{a \code{kalisForwardTable} object, as returned by \code{\link[=MakeForwardTable]{MakeForwardTable()}} and propagated to a target variant by \code{\link[=Forward]{Forward()}}.
This table must be at the same variant location as argument \code{bck}.}

\item{bck}{a \code{kalisBackwardTable} object, as returned by \code{\link[=MakeBackwardTable]{MakeBackwardTable()}} and propagated to a target variant by \code{\link[=Backward]{Backward()}}.
This table must be at the same variant location as argument \code{fwd}.}

\item{pars}{a \code{kalisParameters} object, as returned by \code{\link[=Parameters]{Parameters()}}.}

\item{beta.theta.opts}{a list; see Details in \code{\link[=DistMat]{DistMat()}} documentation page.}

\item{safety.checks}{a logical, should safety checks be applied to the distances?
See \code{\link[=DistMat]{DistMat()}}.}

\item{neighbors}{a logical, should nearest neighbors be pre-calculated?
See \code{\link[=Neighbors]{Neighbors()}}.}

\item{use.forking}{a logical, should forked processes be used?}

\item{forking.chunk.size}{TODO}

\item{mc.preschedule}{TODO}

\item{nthreads}{the number of CPU cores to use.
By default no parallelism is used.}
}
\value{
a \code{kalisClades} object encoding probabilistic clade calls
}
\description{
Utility for calling probabilistic clades at, in between, or excluding variants.
}
\details{
TODO longer description
}
\references{
Christ, R., Wang, X., Aslett, L.J.M., Steinsaltz, D. and Hall, I. (2024) "Clade Distillation for Genome-wide Association Studies." bioRxiv 2024.09.30.615852. Available at: \doi{10.1101/2024.09.30.615852}.
}
