% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheHaplotypes.R
\name{ClearHaplotypeCache}
\alias{ClearHaplotypeCache}
\title{Remove all cached haplotypes and free memory}
\usage{
ClearHaplotypeCache()
}
\value{
Nothing is returned.
}
\description{
Remove all the haplotypes that were cached by a previous caching call and
free up the memory that was allocated for future use.
}
\details{
To achieve higher performance, kalis internally represents haplotypes
in an efficient raw binary format in memory which cannot be directly viewed
or manipulated in R.  In particular, this cache sits outside R's memory
management and will never be garbage collected (unless R is quit or the
package is unloaded).  Therefore, this function is provided to enable freeing
the memory used by this cache.
}
\examples{
# Examples
\dontrun{
ClearHaplotypeCache()
}

}
\seealso{
\code{\link{CacheHaplotypes}} to create a haplotype cache.
}
