% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableCache.R
\name{CreateForwardTableCache}
\alias{CreateForwardTableCache}
\title{Title}
\usage{
CreateForwardTableCache(size = 1, from_recipient = 1, to_recipient = Inf)
}
\arguments{
\item{size}{the maximum amount of RAM (in GB) to devote to this cache.}

\item{from_recipient}{...}

\item{to_recipient}{...}
}
\value{
Return value
}
\description{
Short description
}
\details{
Detailed description
}
\examples{
# Examples
}
\seealso{
See also
}
