% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{DistMat}
\alias{DistMat}
\title{Calculate a Distance Matrix from a Forward Table Object and a Backward Table Object}
\usage{
DistMat(fwd, bck)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}
}
\value{
matrix of distances
}
\description{
Provides an easy function for calculating a local distance matrix.
}
\details{
The forward and backward tables must be at the same locus.
The (i,j)-th element of of the returned matrix is the inferred distance d_(i,j) between haplotypes j copies i at locus fwd$l=bck$l given the haplotypes observed (from locus 1 to L).

d_(i,j) = -( log(p_(i,j)) + log(p_(j,i)) ) / 2 where p_(i,j) is the posterior marginal probability that j copies i at locus fwd$l=bck$l given the haplotypes observed (from locus 1 to L).

By convention, d_(i,i) = 0 for all i.
}
\examples{
DistMat(fwd,bck)
}
\seealso{
\code{\link{PostProb}} to calculate the posterior marginal probabilities p_(i,j);
\code{\link{MakeForward}} to generate Forward table;
\code{\link{Forward}} to propogate a Backward table to a new locus;
\code{\link{MakeBackwardTable}} to generate a Backward table;
\code{\link{Backward}} to propogate a Backward table to a new locus.
}
