% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{DistMat}
\alias{DistMat}
\title{Distance matrix}
\usage{
DistMat(
  fwd,
  bck,
  neg.log = TRUE,
  standardize = FALSE,
  symmetrize = TRUE,
  beta.theta.opts = list(bias = 0.5)
)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{neg.log}{a logical, should a negative logarithm be applied to the posterior copying probabilities to make them distances}

\item{standardize}{a logical; should the columns be centered and scaled to have unit variance (done after neg.log transform)}

\item{symmetrize}{a logical; should the final result be symmetried ( for a matrix M, 0.5*(M + t(M)) )}

\item{beta.theta.opts}{a list; see Details.}
}
\value{
Matrix of distances.
The \eqn{(j,i)}-th element of of the returned matrix is the inferred
distance \eqn{d_{ji}}{d_(j,i)} between haplotypes \eqn{j} and \eqn{i} at
the current locus.
}
\description{
Utility for calculating posterior marginal probabilities and distance matrices at, in between, or excluding variants.
}
\details{
This computes a local probability or distance matrix based on the forward and backward
tables at a certain locus.  The default usage is provide forward and backward tables at the same locus \eqn{l}.
so that the \eqn{(j,i)}-th element of of the returned matrix is the inferred distance
\eqn{d_{ji}}{d_(j,i)} between haplotypes \eqn{j} and \eqn{i} at the current
locus, \eqn{l}, of the two tables given the haplotypes observed (over the
whole sequence).

In particular,

\deqn{d_{ji} = -\frac{log(p_{ji}) + log(p_{ij})}{2}}{d_(j,i) = -( log(p_(j,i)) + log(p_(i,j)) ) / 2}

where \eqn{p_{ji}}{p_(j,i)} is the posterior marginal probability that
\eqn{j} is coped by \eqn{i} at the current locus of the two tables, \eqn{l},
given the haplotypes observed (over the whole sequence).

By convention, \eqn{d_{ii} = 0}{d_(i,i) = 0} for all \eqn{i}.

A matrix of the raw posterior marginal probabilities \eqn{p_{ji}}{p_(j,i)} may be obtained by setting \code{neg.log = FALSE}, \code{standardize = FALSE}, and \code{symmetrize = FALSE}.

This function also allows users to calculate distance matrices in between variants and also to calculate matrices that exclude a set of consecutive variants by passing a
backward table in beta.theta space.  If in beta.theta space, \code{bck$l} may be greater than but not equal to \code{fwd$l}.  When passed a backward table in beta.theta space, the transition probabilities
\code{rho} between \code{fwd$l} and \code{bck$l} will be summed together to form a total, \code{rho_total}.  If \code{bias} is provided as named element in the list \code{beta.theta.opts}, then the tables will be propagated to meet in between with the given \code{bias}.  More explicitly, the forward table will be propagated \code{rho_total * bias} and
the backward table will be propagated a distance of \code{rho_total * (1-bias)}.  The distance to propagate the forward and backward tables respectively can be given by passing \code{rho1step.fwd} and \code{rho1step.bck} as arguments in \code{beta.theta.opts}, which override \code{bias}.
}
\examples{
\dontrun{
# To get the distances at, say, locus 10 ...
Forward(fwd, pars, 10, nthreads = 8)
Backward(bck, pars, 10, nthreads = 8)
d <- DistMat(fwd, bck)
}

}
\seealso{
\code{\link{PostProbs}} to calculate the posterior marginal probabilities
\eqn{p_{ji}}{p_(j,i)};
\code{\link{Forward}} to propogate a Forward table to a new locus;
\code{\link{Backward}} to propogate a Backward table to a new locus.
}
