% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FB.R
\name{Forward}
\alias{Forward}
\title{Propagate an HMM forward table}
\usage{
Forward(fwd, t, Pi, mu, rho, nthreads = 1)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{t}{a locus position to move the forward table to.  Must be greater than
or equal to locus position of table provided in \code{fwd}.}

\item{Pi}{a matrix of background copying probabilities.  Can also provide a
scalar value for uniform background copying probability.}

\item{mu}{a vector of ...}

\item{rho}{a vector of ...}

\item{nthreads}{the number of CPU cores on which to run}
}
\value{
There is nothing returned.  For performance reasons, the forward
  table which was passed in is updated in-place.
}
\description{
Takes a forward table and propagates it in-place to a later locus position.
}
\details{
Detailed description
}
\examples{
fwd <- MakeForwardTable()
Forward(fwd, 100, Pi, mu, rho)
}
\seealso{
\code{\link{MakeForwardTable}} to generate forward table;
  \code{\link{Backward}} for analagous backward induction function.
}
