% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FB.R
\name{Forward}
\alias{Forward}
\title{Propagate a Forward Table}
\usage{
Forward(fwd, pars, t = fwd$l + 1, nthreads = 1)
}
\arguments{
\item{fwd}{\code{kalisForwardTable} as returned by \code{\link{MakeForwardTable}}}

\item{pars}{a \code{kalisParameters} object returned by \code{Parameters}}

\item{t}{a locus position to move the forward table to.  Must be greater than
or equal to \code{fwd$l}.}

\item{nthreads}{the number of CPU cores to use}
}
\value{
There is nothing returned.  For performance reasons, \code{fwd} is updated in-place.
}
\description{
Propagates a \code{kalisForwardTable} to a downstream locus position (nothing is returned: the table is updated in-place).
}
\details{
Forward implements the forward algorithm to advance the rescaled HMM forward probabilities stored in
a \code{kalisForwardTable} \code{fwd} to a new target locus \code{t}.  Note that \code{t} must be greater than the current locus, \code{fwd$l}.
}
\examples{
\dontrun{
fwd <- MakeForwardTable()
Forward(fwd, 100, Pi, mu, rho)
}

}
\seealso{
\code{\link{MakeForwardTable}} to generate a forward table;
\code{\link{Backward}} for the analagous backward propagation function.
}
