% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{ForwardProbs}
\alias{ForwardProbs}
\title{Calculate Normalized Forward Probabilities from a Forward Table Object}
\usage{
ForwardProbs(fwd, log = FALSE)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{log}{logical; if TRUE, the forward probabilities p are returned as log(p).}
}
\value{
matrix of forward probabilities
}
\description{
Provides an easy function for calculating the normalized forward probabilities from a forward table object fwd.
}
\details{
The (j,i)-th element of of the returned matrix is the probability that j is copied by i at locus fwd$l given the haplotypes
observed from locus 1 up through locus fwd$l.  Note: each column represents an independent HMM.

By convention, every diagonal element is zero.
}
\examples{
ForwardProbs(fwd)

}
\seealso{
\code{\link{MakeForwardTable}} to generate forward table;
  \code{\link{Forward}} to propogate a forward table to a new locus.
}
