% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker.R
\name{MakeBackwardTable}
\alias{MakeBackwardTable}
\alias{print.kalisBackwardTable}
\title{Title}
\usage{
MakeBackwardTable(pars, from_recipient = 1, to_recipient = Inf)
}
\arguments{
\item{from_recipient}{...}

\item{to_recipient}{...}
}
\value{
Return value
}
\description{
Short description
}
\details{
Detailed description
}
\examples{
# Examples

}
\seealso{
\code{\link{Backward}} to propagate the newly created table forward
  through the genome.
}
