% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker.R
\name{MakeBackwardTable}
\alias{MakeBackwardTable}
\alias{print.kalisBackwardTable}
\title{Create a \code{kalisBackwardTable}}
\usage{
MakeBackwardTable(pars, from_recipient = 1, to_recipient = Inf)
}
\arguments{
\item{pars}{a \code{kalisParameters} environment specifying the genetics
parameters to be associated with this forward table.  These parameters can
be set up by using the \code{\link{Parameters}} function.}

\item{from_recipient}{first recipient haplotype if creating a partial forward
table.  By default includes from the first recipient haplotype.}

\item{to_recipient}{last recipient haplotype if creating a partial forward
table.  By default includes to the last recipient haplotype.}
}
\value{
a specialized list of class \code{kalisBackwardTable}.  For a given \code{kalisBackwardTable}, \code{bck},
 \code{bck$l} denotes the current locus position of \code{fwd}.  \code{bck$beta} is a matrix of rescaled backward probabilities under the Li \& Stephens HMM.
Each column of \code{bck$beta} corresponds to an independent HMM such that \eqn{\beta^\ell_{ji}}
is proportional to the probability of observing haplotype \eqn{i} from locus \eqn{l+1} up through locus \eqn{L} given that haplotype \eqn{j}
is copied by haplotype \eqn{i} at locus \eqn{\ell}.  \code{bck$beta.g} and \code{bck$beta.g2}
are both vectors containing scaling constants needed to continue propagating the HMM (please see kalis paper for details).

\code{kalisBackwardTable} also carries with it a checksum key for the parameters \code{pars} it was provided.
If a user attempts to interact a \code{kalisBackwardTable} with a \code{kalisForwardTable} with mismatched
parameters, an error will be thrown.
}
\description{
Allocates the memory for and initializes up a backward table.
}
\details{
\code{MakeBackwardTable} returns a \code{kalisBackwardTable} object appropriate for
a given set of haplotypes (that must have already been cached by \code{CacheAllHaplotypes}) and
a given set of HMM parameters specified by \code{pars}.  The returned \code{kalisBackwardTable} is
initialized at the end of the cached haplotypes (technically \code{bck$l}=2,147,483,647 for computational reasons)
and is ready to be propagated to a given target locus with the function \code{\link{Backward}}.

Since there is an independent hidden Markov model run for each recipient
haplotype, it is possible to create a partial forward table object which
corresponds to a subset of recipients using the \code{from_recipient} and
\code{to_recipient} arguments.
}
\examples{
\dontrun{
MakeBackwardTable(...)
}

}
\seealso{
\code{\link{Backward}} to propagate the newly created \code{kalisBackwardTable}.
}
