% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{Parameters}
\alias{Parameters}
\title{Create rho from standard parameters}
\usage{
Parameters(rho = rep(0, get("L", envir = pkgVars) - 1), mu = 1e-08,
  Pi = NULL, check.rho = TRUE)
}
\arguments{
\item{rho}{recombination probability vector (must be L-1 long), see \code{\link{CalcRho}}}

\item{mu}{a scalar (for uniform) or vector (for varying) mutation costs.}

\item{Pi}{leaving the default of uniform copying probabilities is recommended for
computational efficiency.  If desired, a full matrix of background copying
probabilities can be provided, such that the (j,i)-th element is the background
probability that j is copied by i.  Hence, (a) the diagonal must be zero; and (b)
the columns of Pi must sum to 1.}

\item{check.rho}{if TRUE, a check that rho is within machine precision will be
performed.  If you have created rho using \code{\link{CalcRho}} with \code{floor=TRUE}
then this will be satisfied.}
}
\value{
A \code{kalisParameters} object.
}
\description{
Convenience function for calculating kalis recombination (renewal) probabilities from standard Pop. Gen. parameters.
}
\details{
Detailed description
}
\examples{
\dontrun{
pars <- Parameters(...)
}

}
\seealso{
\code{\link{MakeForwardTable}}, \code{\link{MakeForwardTable}} which
construct table objects which internally reference a parameters environment.
}
