% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{Parameters}
\alias{Parameters}
\title{Construct parameter specification}
\usage{
Parameters(
  rho = rep(0, get("L", envir = pkgVars) - 1),
  mu = 1e-08,
  Pi = NULL,
  use.speidel = FALSE,
  check.rho = TRUE
)
}
\arguments{
\item{rho}{recombination probability vector (must be L-1 long).
See \code{\link{CalcRho}} for assistance constructing this from standard
genetics parameters.}

\item{mu}{a scalar (for uniform) or vector (for varying) mutation costs.}

\item{Pi}{leaving the default of uniform copying probabilities is recommended
for computational efficiency.
If desired, a full matrix of background copying probabilities can be
provided, such that the (j,i)-th element is the background probability that
j is copied by i.
Hence, (a) the diagonal must be zero; and (b) the columns of Pi must sum to
1.}

\item{check.rho}{if \code{TRUE}, a check that rho is within machine precision
will be performed.
If you have created rho using \code{\link{CalcRho}} with \code{floor=TRUE}
then this will be satisfied automatically.}
}
\value{
A \code{kalisParameters} object, suitable for use to create the
standard forward and backward recursion tables with
\code{\link{MakeForwardTable}} and \code{\link{MakeBackwardTable}}.
Note you will also need to provide this parameters object when propagating
those tables using either \code{\link{Forward}} or \code{\link{Backward}}.
}
\description{
Specify a parameter set to be used for a particular Li and Stephens hidden
Markov model run.
}
\details{
There are 3 parameters which must be specified before the forward or backward
equations in the Li and Stephens hidden Markov model can be run.  These are
the vector of recombination probabilities, the mutation costs, and the
prior copying probabilities.

\strong{Recombination probabilities, \code{rho}}

This is a vector parameter which must have length \code{L-1}, where \code{L}
is the length (number of loci) of the haplotype sequences which have been
loaded into the memory cache (using \code{\link{CacheHaplotypes}}).
Note that element i of this vector should be the recombination probability
between locus \code{i} and \code{i+1}.

There is a utility function, \code{\link{CalcRho}}, to assist with creating
this recombination vector based on the more standard genetic parameters of
Morgan distances, effective population size and a scalar power.

By default, the recombination probabilities are set to zero everywhere.

\strong{Mutation costs, \code{mu}}

The mutation costs may be specified either as uniform across the whole length
of the haplotypes (by providing a single scalar value), or may be varying
at each locus (by providing a vector of length equal to the length of the
haplotype sequences which have been loaded into the memory cache).

By default, mutation costs are set to \eqn{10^{-8}}{10^-8}.

\strong{Copying probabilities, \code{Pi}}

The original Li and Stephens model assumed that each haplotype has an equal
prior probability of copying from any other.
However, in the spirit of ChromoPainter (Lawson et al., 2012) we allow a
matrix of prior copying probabilities.

The copying probabilities may be specified as a matrix of size \eqn{N \times N}{N * N}
(where \eqn{N} is the number of haplotypes which have been loaded into the
memory cache), or for uniform copying probabilities need not be specified
(resulting in copying probability \eqn{\frac{1}{N-1}}{1/(N-1)} everywhere).
Note that the diagonal must by definition be zero and columns must sum to
one.

Note that there is a computational cost associated with non-uniform copying
probabilities, so absent a strong motivation it is recommended to leave the
default of uniform probabilities (NB \emph{do not} specify a uniform matrix which
would incur a computation cost too).
}
\examples{
\dontrun{
# To use all the defaults
pars <- Parameters()

# To use indirect genetics parameters which you have already specified in
# morgan.dist, Ne and gamma variables to specify rho, leaving mu and Pi as
# defaults
pars <- Parameters(CalcRho(morgan.dist, Ne, gamma))

# Then use to create a table (eg for forward) ...
fwd <- MakeForwardTable(pars)
# ... and to propagate it
Forward(fwd, pars, 10, nthreads = 8)
}

}
\references{
Lawson, D. J., Hellenthal, G., Myers, S., & Falush, D. (2012). Inference of
population structure using dense haplotype data. \emph{PLoS genetics}, \strong{8}(1).
}
\seealso{
\code{\link{MakeForwardTable}} and \code{\link{MakeForwardTable}}
which construct table objects which internally reference a parameters
environment.
\code{\link{Forward}} and \code{\link{Backward}} which propagate those
tables according to the Li and Stephens model.
}
