% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{Parameters}
\alias{Parameters}
\title{Setup genetics parameters}
\usage{
Parameters(morgan.dist = 0, Ne = 1, gamma = 1, mu = 1e-08,
  Pi = NULL)
}
\arguments{
\item{morgan.dist}{a vector of recombination distances between loci, in Morgans.
Note element i of this vector should be the distance between loci i and i+1
(not i and i-1), and thus length one less than the haplotype length.}

\item{Ne}{a scalar for the effective population size.  Can be autotuned, see ...}

\item{gamma}{a scalar power to which the Morgan distances are raised.  Can be
autotuned, see ...}

\item{mu}{a scalar (for uniform) or vector (for varying) mutation costs.}

\item{Pi}{leaving the default of uniform copying probabilities is recommended for
computational efficiency.  If desired, a full matrix of background copying
probabilities can be provided, such that the (i,j)-th element is the background
probability that j copies i.  Hence, (a) the diagonal must be zero; and (b)
the columns of Pi must sum to 1.}
}
\value{
A new parameters environment.
}
\description{
Sets up the genetics parameters to be used for a problem.
}
\details{
Detailed description
}
\examples{
\dontrun{
pars <- Parameters(...)
}

}
\seealso{
\code{\link{MakeForwardTable}}, \code{\link{MakeForwardTable}} which
  construct table objects which internally reference a parameters environment.
}
