% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{plot.kalisDistanceMatrix}
\alias{plot.kalisDistanceMatrix}
\title{Plotting function for a kalisDistanceMatrix object}
\usage{
\method{plot}{kalisDistanceMatrix}(x, ...)
}
\arguments{
\item{d}{a kalisDistanceMatrix}
}
\value{
There is nothing returned.
}
\description{
Clusters the given distance matrix and generates a heatmap to display it.
}
