% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{PostProbs}
\alias{PostProbs}
\title{Posterior marginal probabilities}
\usage{
PostProbs(fwd, bck, log = FALSE)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{log}{logical; defaults to \code{FALSE}.
If \code{TRUE}, the logarithm of the posterior marginal probabilities are
returned.}
}
\value{
Matrix of posterior marginal probabilities.
The \eqn{(j,i)}-th element of of the returned matrix is the probability
that \eqn{j} is copied by \eqn{i} at the current locus, \eqn{l}, of the two
tables, given the haplotypes observed (over the whole sequence).
}
\description{
Calculated the posterior marginal probabilities at a given locus using
forward and backward tables propagated to that position.
}
\details{
The forward and backward tables must be at the same locus in order for them
to be combined to yield the posterior marginal probabilities at locus
\eqn{l}.
The \eqn{(j,i)}-th element of of the returned matrix is the probability that
\eqn{j} is copied by \eqn{i} at the current locus, \eqn{l}, of the two
tables, given the haplotypes observed (over the whole sequence).

Note that each column represents an independent HMM.

By convention, every diagonal element is zero.
}
\examples{
\dontrun{
# To get the posterior probabilities at, say, locus 10 ...
Forward(fwd, pars, 10, nthreads = 8)
Backward(bck, pars, 10, nthreads = 8)
p <- PostProbs(fwd, bck)
}

}
\seealso{
\code{\link{DistMat}} to generate calculate \eqn{d_{ji}}{d_(j,i)} distances
directly;
\code{\link{Forward}} to propogate a Forward table to a new locus;
\code{\link{Backward}} to propogate a Backward table to a new locus.
}
