% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probs.R
\name{PostProbs}
\alias{PostProbs}
\title{Calculate Posterior Marginal Probabilities from a Forward Table Object and a Backward Table Object}
\usage{
PostProbs(fwd, bck, log = TRUE)
}
\arguments{
\item{fwd}{a forward table as returned by \code{\link{MakeForwardTable}}}

\item{bck}{a backward table as returned by \code{\link{MakeBackwardTable}}}

\item{log}{logical; if TRUE (default), the posterior marginal probabilities p are returned as log(p).}
}
\value{
matrix of posterior marginal probabilities
}
\description{
Provides an easy function for calculating the log posterior marginal probabilities.
}
\details{
Note that the forward and backward tables must be at the same locus in order for them to be combined to yield the posterior marginal probabilities.
The (i,j)-th element of of the returned matrix is the probability that j copies i at locus fwd$l=bck$l given the haplotypes observed (from locus 1 to L).

Since a haplotype is not allowed to copy from itself, every diagonal element is zero.
}
\examples{
fwd <- MakeForwardTable()
Forward(fwd, 100, Pi, mu, rho)
bck <- MakeBackwardTable()
Backward(bck, 100, Pi, mu, rho)
BackwardProbs(bck)
PostProbs(fwd,bck)
}
\seealso{
\code{\link{MakeForward}} to generate Forward table;
  \code{\link{Forward}} to propogate a Backward table to a new locus;
\code{\link{MakeBackwardTable}} to generate a Backward table;
  \code{\link{Backward}} to propogate a Backward table to a new locus.
}
