% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clades.R
\name{PruneCladeMat}
\alias{PruneCladeMat}
\title{Prune called sprigs or singletons from inferred clade matrix}
\usage{
PruneCladeMat(M, neigh, sprigs, prune = "singleton.info", from_recipient = 1L)
}
\arguments{
\item{M}{clade matrix such as that updated by \code{\link[=CladeMat]{CladeMat()}}}

\item{neigh}{a list of nearest neighborhoods as retuned by \code{\link[=CladeMat]{CladeMat()}}}

\item{sprigs}{a sprigs object as returned by \code{\link[=Sprigs]{Sprigs()}}}

\item{prune}{a character indicating the type of information to be removed from the \code{\link[=CladeMat]{CladeMat()}}. See Details.}

\item{from_recipient}{haplotype index at which to start trace calculation --- useful for distributed computation (experimental feature, more documentation to come\if{html}{\out{<!-- TODO -->}})}
}
\value{
There is nothing returned.

\strong{NOTE:} for performance reasons, \code{M} is updated in-place.
}
\description{
Use haplotype nearest neighborhoods and other information returned by \code{\link[=CladeMat]{CladeMat()}} to efficiently remove structure corresponding to singletons or called sprigs from clade matrix 'M'
}
\examples{
# TODO

}
