% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheAllSequences.R
\name{QueryCache}
\alias{QueryCache}
\title{Retrieve haplotypes from memory cache}
\usage{
QueryCache(ids = NA, start = 1, length = NA)
}
\arguments{
\item{ids}{which haplotypes to retrieve from the cache}

\item{start}{the first locus position to retrieve for the specified haplotypes}

\item{length}{the number of consecutive loci to return from the start position}
}
\value{
A matrix of 0/1 integers with \code{length} rows and
  \code{length(ids)} columns, such that haplotypes appear in columns.
}
\description{
Retrieve haplotypes from the memory cache, converting the raw binary into
a simple integer 0/1 vector for use in R.
}
\details{
To achieve higher performance, kalis internally represents haplotype sequences
in an efficient raw binary format in memory which cannot be directly viewed
or manipulated in R.  This function enables users to copy whole or partial
haplotype sequences out of this low-level format and into a standard R
vector of 0's and 1's.
}
\examples{
# Examples
\dontrun{
}

}
\seealso{
\code{\link{Forward}} to propagate the newly created table forward
  through the genome.
}
