% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMaker.R
\name{ResetTable}
\alias{ResetTable}
\title{Reset Forward/Backward tables}
\usage{
ResetTable(tbl)
}
\arguments{
\item{tbl}{a \code{kalisForwardTable} or \code{kalisBackwardTable} object
which is to be reset.}
}
\description{
Resets the provided forward/backward table so it can propagate from the start/end of the Li and Stephens hidden Markov model.
}
\details{
It is \emph{much} faster to reset a forward/backward table rather than remove and make a new one.
This function marks a table as reset so that it will be propagated as if freshly allocated.
}
\examples{
# Load the toy haplotype example and set toy parameters
data("SmallHaps")
data("SmallMap")

CacheHaplotypes(SmallHaps)

rho <- CalcRho(diff(SmallMap))
pars <- Parameters(rho)

# Create the forward table we want to propagate
fwd <- MakeForwardTable(pars)

# Move to variant 10
Forward(fwd, pars, 10)
fwd

# This does **NOT** work as intended, cannot reverse from variant 10:
try(Forward(fwd, pars, 5))

# Could reset the table though:
ResetTable(fwd)
Forward(fwd, pars, 5)
fwd

}
\references{
Aslett, L.J.M. and Christ, R.R. (2024) "kalis: a modern implementation of the Li & Stephens model for local ancestry inference in R", \emph{BMC Bioinformatics}, \strong{25}(1). Available at: \doi{10.1186/s12859-024-05688-8}.
}
\seealso{
\code{\link[=MakeForwardTable]{MakeForwardTable()}}, \code{\link[=MakeBackwardTable]{MakeBackwardTable()}} to create fresh tables.
}
