% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallHaps-data.R
\docType{data}
\name{SmallHaps}
\alias{SmallHaps}
\title{SmallHaps Example Phased Haplotype Dataset for kalis}
\format{An object of class \code{matrix} where each column is a
simulated haplotype.
This is suitable for passing directly to \code{\link{CacheHaplotypes}}.}
\usage{
data(SmallHaps)
}
\description{
Simulated dataset with N=100 haplotypes at L=500 loci generated using msprime
(Kelleher et al., 2016).
}
\examples{
\dontrun{
data(SmallHaps)
# Plot Allele Frequencies
hist(rowMeans(SmallHaps),breaks=20)

# Store in kalis hdf5 format
WriteIndividualHaplotypeH5("SmallHaps.h5", SmallHaps)

# Import into kalis cache directly ...
CacheHaplotypes(SmallHaps)
# ... or via the HDF5 file written out above
CacheHaplotypes("SmallHaps.h5")
}

}
\references{
Kelleher, J., Etheridge, A. M., & McVean, G. (2016). Efficient coalescent
simulation and genealogical analysis for large sample sizes.
\emph{PLoS computational biology}, \strong{12}(5).
}
\keyword{datasets}
