% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clades.R
\name{Sprigs}
\alias{Sprigs}
\title{Title TODO}
\usage{
Sprigs(x, old.sprigs = FALSE)
}
\arguments{
\item{x}{TODO}

\item{old.sprigs}{TODO}
}
\value{
TODO
}
\description{
Short Description TODO
}
\details{
Long Description TODO
}
\examples{
# TODO

}
